seo_agent/app/config.py
# app/config.py
import os
from dotenv import load_dotenv

load_dotenv()

GEMINI_API_KEY: str = os.getenv("GEMINI_API_KEY", "")
SERP_API_KEY: str = os.getenv("SERP_API_KEY", "")
GEMINI_MODEL: str = os.getenv("GEMINI_MODEL", "gemini-2.0-flash")
API_KEY: str = os.getenv("API_KEY", "")
