seo_agent/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
requests==2.31.0
beautifulsoup4==4.12.3
pydantic==2.6.4
python-dotenv==1.0.1
google-genai>=1.0.0
